+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet!
+### Fixed
+
+- Add support for sorting `class` attribute on custom elements and Astro components ([#129](https://github.com/tailwindlabs/prettier-plugin-tailwindcss/pull/129))
 
 ## [0.2.3] - 2023-02-15